// src/app/mytypes/types.ts
  export interface TypeAnnonce {
    id: string; // Identifiant unique
    name: string; // Nom du type d'annonce
    nameAr: string; // Nom du type d'annonce en arabe
    priority: number; // Priorité du type d'annonce
    createdAt: Date; // Date de création du type d'annonce
  }
  
  export interface SubCategory {
    id: number;
    name: string;
    nameAr: string;
    priority?: number;
    createdAt: Date;
  }
  
  // Interface pour la table Categories
  export interface Category {
    id: string; // Identifiant unique
    typeAnnonceid: string; // Identifiant du type d'annonce (clé étrangère)
    name: string;
    nameAr: string; // Nom de la catégorie
    priority: number; // Priorité de la catégorie
    createdAt: Date; // Date de création de la catégorie
  }